Improved a little bit how the session works
